StudentSorter/src/main/java/org/tpg/models/Student.java
package org.tpg.models;

public class Student {
	private long id;
	private String name;
	private double gpa;
	
	public Student() {
	}
	
	public Student(int id, String name, double gpa) {
		this.id = id;
		this.name = name;
		this.gpa = gpa;
	}
	
	public String toString() {
		return this.id + " " + this.name;
	}
	
	public long getId() {
		return id;
	}
	public void setId(long id) {
		this.id = id;
	}
	public String getName() {
		return name;
	}
	public void setName(String name) {
		this.name = name;
	}
	public double getGpa() {
		return gpa;
	}
	public void setGpa(double gpa) {
		this.gpa = gpa;
	}
}
